pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-research-agent"
version = "0.1.0"
description = "A modular deep research pipeline with planning, retrieval, synthesis, and claim-level verification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "anthropic>=0.40.0",
  "tavily-python>=0.3.0",
  "langgraph>=1.0.0",
  "python-dotenv>=1.0.0",
  "pandas>=2.0.0",
  "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
